refactor(*.*): 사용하지 않는 파일 제
- jpa를 사용하면서 테스트 수정 된 부분을 적용
